LPS-137441 - Move from frontend-js to frontend-icons
