+++ modules/apps/frontend-icons/frontend-icons-contributor-extender/build.gradle
