+++ modules/apps/frontend-icons/frontend-icons-contributor-extender/src/main/java/com/liferay/frontend/icons/contributor/extender/IconResourcesContributor.java
@@ -12,29 +12,15 @@
  * details.
  */
 
-package com.liferay.frontend.js.icon.contributor.extender.internal;
+package com.liferay.frontend.icons.contributor.extender;
 
-import com.liferay.frontend.js.icon.contributor.extender.IconResource;
-import com.liferay.frontend.js.icon.contributor.extender.IconResourcesContributor;
-
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
 
 /**
  * @author Iván Zaera Avellón
  */
-public class IconResourcesContributorImpl implements IconResourcesContributor {
+public interface IconResourcesContributor {
 
-	public void addIconResource(IconResource iconResource) {
-		_iconResources.add(iconResource);
-	}
-
-	@Override
-	public Collection<IconResource> getIconResources() {
-		return _iconResources;
-	}
-
-	private final List<IconResource> _iconResources = new ArrayList<>();
+	public Collection<IconResource> getIconResources();
 
 }
