--- modules/apps/frontend-js/frontend-js-icon-contributor-extender/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Frontend JS Icon Contributor Extender
-Bundle-SymbolicName: com.liferay.frontend.js.icon.contributor.extender
-Bundle-Version: 3.0.0
